+++ pkg/service/factory_build.go
@@ -6,13 +6,16 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"time"
 
 	"github.com/google/uuid"
 	"github.com/jonboulle/clockwork"
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/cli/dashboardrender"
 	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
 	"github.com/portpowered/infinite-you/pkg/factory"
 	factoryevents "github.com/portpowered/infinite-you/pkg/factory/events"
+	"github.com/portpowered/infinite-you/pkg/factory/projections"
 	"github.com/portpowered/infinite-you/pkg/factory/runtime"
 	"github.com/portpowered/infinite-you/pkg/factory/state"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
@@ -361,4 +364,64 @@ func providerCommandRunnerForMode(cfg *FactoryServiceConfig, runtimeCfg interfac
 		RuntimeConfig: runtimeCfg,
 		Next:          cfg.ProviderCommandRunnerOverride,
 	}
+}
+
+func (fs *FactoryService) dashboardLoop(ctx context.Context) {
+	ticker := time.NewTicker(30 * time.Second)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+			fs.renderDashboard(ctx)
+		}
+	}
+}
+
+func (fs *FactoryService) renderDashboard(ctx context.Context) {
+	now := factory.EnsureClock(fs.clock).Now()
+	input, err := fs.buildSimpleDashboardRenderInput(ctx, now)
+	if err != nil {
+		if fs.logger != nil {
+			fs.logger.Error("simple dashboard render failed", zap.Error(err))
+		}
+		return
+	}
+	fs.cfg.SimpleDashboardRenderer(input)
+}
+
+func (fs *FactoryService) buildSimpleDashboardRenderInput(ctx context.Context, now time.Time) (SimpleDashboardRenderInput, error) {
+	es, err := fs.GetEngineStateSnapshot(ctx)
+	if err != nil {
+		return SimpleDashboardRenderInput{}, err
+	}
+	renderData, err := fs.simpleDashboardRenderData(ctx, es.TickCount, es.ActiveThrottlePauses)
+	if err != nil {
+		return SimpleDashboardRenderInput{}, err
+	}
+	return SimpleDashboardRenderInput{
+		EngineState: *es,
+		RenderData:  renderData,
+		Now:         now,
+	}, nil
+}
+
+func (fs *FactoryService) simpleDashboardRenderData(
+	ctx context.Context,
+	selectedTick int,
+	activeThrottlePauses []interfaces.ActiveThrottlePause,
+) (dashboardrender.SimpleDashboardRenderData, error) {
+	events, err := fs.GetFactoryEvents(ctx)
+	if err != nil {
+		return dashboardrender.SimpleDashboardRenderData{}, err
+	}
+	worldState, err := projections.ReconstructFactoryWorldState(events, selectedTick)
+	if err != nil {
+		return dashboardrender.SimpleDashboardRenderData{}, err
+	}
+	renderData := dashboardrender.SimpleDashboardRenderDataFromWorldState(worldState)
+	renderData.ActiveThrottlePauses = projections.ProjectActiveThrottlePauses(worldState.Topology, activeThrottlePauses)
+	return renderData, nil
 }
