--- pkg/service/factory_runtime_support.go
@@ -1,191 +1,0 @@
-package service
-
-import (
-	"errors"
-	"fmt"
-	"path/filepath"
-	"strings"
-
-	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
-	"github.com/portpowered/infinite-you/pkg/factory"
-	factory_context "github.com/portpowered/infinite-you/pkg/factory/context"
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-	"github.com/portpowered/infinite-you/pkg/replay"
-	"go.uber.org/zap"
-)
-
-func validateReplayModeConfig(cfg *FactoryServiceConfig) error {
-	if cfg == nil {
-		return fmt.Errorf("factory service config is required")
-	}
-	if cfg.RecordPath != "" && cfg.ReplayPath != "" {
-		return fmt.Errorf("--record and --replay cannot be used together")
-	}
-	return nil
-}
-
-func loadFactoryConfigForMode(cfg *FactoryServiceConfig) (*factoryconfig.LoadedFactoryConfig, *interfaces.ReplayArtifact, error) {
-	if cfg.ReplayPath == "" {
-		loaded, err := factoryconfig.LoadRuntimeConfig(cfg.Dir, cfg.WorkstationLoader)
-		if loaded != nil {
-			loaded.SetRuntimeBaseDir(cfg.ExecutionBaseDir)
-		}
-		return loaded, nil, err
-	}
-	artifact, err := replay.Load(cfg.ReplayPath)
-	if err != nil {
-		return nil, nil, fmt.Errorf("load replay artifact: %w", err)
-	}
-	runtimeCfg, err := replay.RuntimeConfigFromGeneratedFactory(artifact.Factory)
-	if err != nil {
-		return nil, nil, fmt.Errorf("load embedded replay config: %w", err)
-	}
-	loaded, err := factoryconfig.NewLoadedFactoryConfig(runtimeCfg.FactoryDir(), runtimeCfg.Factory, runtimeCfg)
-	if err != nil {
-		return nil, nil, fmt.Errorf("build embedded replay config: %w", err)
-	}
-	loaded.SetRuntimeBaseDir(cfg.ExecutionBaseDir)
-	return loaded, artifact, nil
-}
-
-func warnReplayMetadataMismatches(cfg *FactoryServiceConfig, artifact *interfaces.ReplayArtifact, logger *zap.Logger) {
-	if artifact == nil || cfg == nil || cfg.Dir == "" {
-		return
-	}
-	current, err := factoryconfig.LoadRuntimeConfig(cfg.Dir, cfg.WorkstationLoader)
-	if err != nil {
-		return
-	}
-	currentFactory, err := replay.GeneratedFactoryFromRuntimeConfig(
-		current.FactoryDir(),
-		current.FactoryConfig(),
-		current,
-		replay.WithGeneratedFactorySourceDirectory(current.FactoryDir()),
-		replay.WithGeneratedFactoryWorkflowID(cfg.WorkflowID),
-	)
-	if err != nil {
-		return
-	}
-	for _, warning := range replay.FactoryMetadataWarnings(artifact.Factory, currentFactory) {
-		logger.Warn("replay artifact metadata differs from current checkout",
-			zap.String("category", replay.DivergenceCategoryConfigMismatch),
-			zap.String("metadata_key", warning.Key),
-			zap.String("artifact", warning.Artifact),
-			zap.String("current", warning.Current),
-		)
-	}
-}
-
-func warnPortableBundledReplacementReport(
-	logger *zap.Logger,
-	message string,
-	replacements []factoryconfig.PortableBundledFileReplacement,
-) {
-	if logger == nil || len(replacements) == 0 {
-		return
-	}
-	targets := make([]string, 0, len(replacements))
-	for _, replacement := range replacements {
-		targets = append(targets, replacement.TargetPath)
-	}
-	logger.Warn(message, zap.Strings("target_paths", targets))
-}
-
-func runtimeWorkflowContext(cfg *interfaces.FactoryConfig) *factory_context.FactoryContext {
-	projectID := factory_context.DefaultProjectID
-	if cfg != nil && cfg.Project != "" {
-		projectID = factory_context.ResolveProjectID(cfg.Project, nil, nil)
-	}
-	return &factory_context.FactoryContext{
-		ProjectID: projectID,
-		EnvVars:   make(map[string]string),
-	}
-}
-
-func newRecordingArtifact(
-	cfg *FactoryServiceConfig,
-	factoryDir string,
-	factoryCfg *interfaces.FactoryConfig,
-	runtimeCfg interfaces.RuntimeDefinitionLookup,
-	clock factory.Clock,
-) (*interfaces.ReplayArtifact, error) {
-	if cfg.RecordPath == "" {
-		return nil, nil
-	}
-	now := factory.EnsureClock(clock).Now().UTC()
-	generatedFactory, err := replay.GeneratedFactoryFromRuntimeConfig(
-		factoryDir,
-		factoryCfg,
-		runtimeCfg,
-		replay.WithGeneratedFactorySourceDirectory(factoryDir),
-		replay.WithGeneratedFactoryWorkflowID(cfg.WorkflowID),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("build replay artifact config: %w", err)
-	}
-	return replay.NewEventLogArtifactFromFactory(now, generatedFactory, &interfaces.ReplayWallClockMetadata{
-		StartedAt: now,
-	}, interfaces.ReplayDiagnostics{})
-}
-
-func (fs *FactoryService) finalizeRuntimeArtifacts(runtimeBundle *replacementFactoryRuntime) error {
-	if runtimeBundle == nil {
-		return nil
-	}
-	var errs []error
-	if runtimeBundle.recording != nil {
-		runtimeBundle.recording.Finish(factory.EnsureClock(fs.clock).Now().UTC())
-		if err := runtimeBundle.recording.Flush(); err != nil {
-			errs = append(errs, err)
-		}
-		if err := runtimeBundle.recording.Err(); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	if runtimeBundle.logSink != nil {
-		if err := runtimeBundle.logSink.Close(); err != nil {
-			errs = append(errs, err)
-		}
-	}
-	return errors.Join(errs...)
-}
-
-func newSessionLogger(base *zap.Logger, sessionID string, folderPath string, factoryDir string) *zap.Logger {
-	if base == nil {
-		base = zap.NewNop()
-	}
-	return base.With(
-		zap.String("session_id", sessionID),
-		zap.String("folder_path", folderPath),
-		zap.String("factory_dir", factoryDir),
-	)
-}
-
-func sessionScopedRecordPath(basePath string, sessionID string) string {
-	if strings.TrimSpace(basePath) == "" {
-		return basePath
-	}
-	if strings.Contains(basePath, "__factory_session_id__") {
-		return strings.ReplaceAll(basePath, "__factory_session_id__", sessionID)
-	}
-	if sessionID == defaultFactorySessionID {
-		return basePath
-	}
-	ext := filepath.Ext(basePath)
-	base := strings.TrimSuffix(basePath, ext)
-	return base + "." + sessionID + ext
-}
-
-func (r *replacementFactoryRuntime) runtimeLogger() *zap.Logger {
-	if r == nil || r.logger == nil {
-		return zap.NewNop()
-	}
-	return r.logger
-}
-
-func runtimeModeOrDefault(mode interfaces.RuntimeMode) interfaces.RuntimeMode {
-	if mode == "" {
-		return interfaces.RuntimeModeBatch
-	}
-	return mode
-}
