--- pkg/service/factory_service_mode_startup.go
@@ -1,53 +1,0 @@
-package service
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"time"
-)
-
-func (fs *FactoryService) waitForServiceModeStartupWorkReadability(ctx context.Context, serviceMode bool) error {
-	if !serviceMode || fs.cfg.WorkFile == "" || fs.cfg.APIServerReady == nil || fs.cfg.Port <= 0 || fs.cfg.APIServerStarter == nil {
-		return nil
-	}
-	apiServerExit := fs.apiServerExit
-	select {
-	case <-fs.cfg.APIServerReady:
-	case err := <-apiServerExit:
-		return startupReadinessError(err)
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-
-	timer := time.NewTimer(serviceModeStartupWorkReadabilityDelay)
-	defer timer.Stop()
-
-	select {
-	case <-timer.C:
-		return nil
-	case err := <-apiServerExit:
-		return startupReadinessError(err)
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-func (fs *FactoryService) failServiceModeStartup(currentRuntime *liveRuntimeHandle, startupErr error) error {
-	fs.clearRunState()
-	fs.unregisterLiveSession(defaultFactorySessionID)
-	if currentRuntime == nil {
-		return startupErr
-	}
-	if stopErr := fs.stopLiveRuntime(currentRuntime); stopErr != nil && !errors.Is(stopErr, context.Canceled) {
-		return errors.Join(startupErr, stopErr)
-	}
-	return startupErr
-}
-
-func startupReadinessError(err error) error {
-	if err == nil {
-		return fmt.Errorf("wait for service-mode startup work readiness: API server stopped before signaling readiness")
-	}
-	return fmt.Errorf("wait for service-mode startup work readiness: %w", err)
-}
